chore: refactor integration actions into it's own component; fixes #196
