+++ app/ui-react/syndesis/src/modules/integrations/components/index.ts
@@ -4,3 +4,4 @@ export * from './IntegrationDetailSteps';
 export * from './IntegrationEditorBreadcrumbs';
 export * from './IntegrationEditorSidebar';
 export * from './IntegrationEditorStepAdder';
+export * from './WithIntegrationActions';
